pods/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
psutil>=5.9.0
requests>=2.31.0
python-dotenv>=1.0.0
kubernetes>=29.0.0
pyyaml>=6.0.1
